:sparkles: (getStandardChartPointsFromVersion) compute the new standard based on ressources capbilities
At the beginning of sprint, if the relase respect the recommandations (time and volume), then the standard should be equal to the forecast).
Thus the project team can see problems if the team doesn't deliver one day. It avoids to say "I know we are late, but it is normal, Guillaume was missing yesterday, and the following days we will make up the late